requirements.txt
Flask==3.0.3
PyMuPDF==1.24.1
google-generativeai==1.45.0